requirements.txt
docopt==0.6.2
requests==2.6.0
python3-pika==0.9.14
lxml==3.4.4
cssselect==0.9.1
